+++ config/zeus-wind.php
@@ -11,4 +11,14 @@ return [
         'disk' => 'public',
         'directory' => 'logos',
     ],
+
+    'title' => config('app.name', 'Laravel').' | '.'Contact Us',
+
+    'description' => config('app.name', 'Laravel').' Contact Us',
+
+    'color' => '#F5F5F4',
+
+    'default_status' => 'NEW',
+
+
 ];