+++ resources/views/contact.blade.php
@@ -1,7 +1,4 @@
 <div>
-    <x-slot name="title">
-        {{ __('Contact us') }}
-    </x-slot>
     <x-slot name="header">
         <h2>{{ __('Contact us') }}</h2>
     </x-slot>
@@ -16,13 +13,10 @@
         <form wire:submit.prevent="store">
             <div class="max-w-4xl mx-auto my-4">
                 {{ $this->form }}
-
                 <div class="p-4 text-center">
                     <x-zeus::button type="submit">{{ __('Send') }}</x-zeus::button>
                 </div>
             </div>
-
         </form>
     @endif
-
 </div>