+++ src/Filament/Resources/DepartmentResource.php
@@ -24,6 +24,11 @@ class DepartmentResource extends Resource
     protected static ?int $navigationSort = 1;
     protected static ?string $navigationGroup = 'Wind';
 
+    protected static function shouldRegisterNavigation() : bool
+    {
+        return config('zeus-wind.enableDepartments');
+    }
+
     public static function form(Form $form): Form
     {
         return $form