+++ src/Filament/Resources/DepartmentResource/Pages/CreateDepartment.php
@@ -8,4 +8,9 @@ use LaraZeus\Wind\Filament\Resources\DepartmentResource;
 class CreateDepartment extends CreateRecord
 {
     protected static string $resource = DepartmentResource::class;
+
+    public function mount(): void
+    {
+        abort_unless(config('zeus-wind.enableDepartments'), 404);
+    }
 }