+++ src/Filament/Resources/DepartmentResource/Pages/EditDepartment.php
@@ -8,4 +8,9 @@ use LaraZeus\Wind\Filament\Resources\DepartmentResource;
 class EditDepartment extends EditRecord
 {
     protected static string $resource = DepartmentResource::class;
+
+    public function mount($record): void
+    {
+        abort_unless(config('zeus-wind.enableDepartments'), 404);
+    }
 }